This is a very important fix update:
The RC Controller sets different limits for the radio_in[sheet_output] depending on the winch. When we put in a new winch, the RC values are different for when the sheet is fully out vs. fully in:

Sheet Condition vs. RC Value (located in the variable radio_in[sheet_output])
(fully out sheet % = 0) sheet_end = 1932
(fully in sheet % = 100) sheet_fully_in = 1342

In RC mode, the arduino sails with the direct output from the RC and outputs it to the Servo. So, the RC gives 1932 when you tell the sheet to go fully out (by moving the left stick all the way up), and sends this directly to the winch servo.

In Auto mode, the arduino receives a sheetPercentage, calculated by the pi. It takes this sheet percentage, and calculates what to send to the winch, via this equation:

outputToWinch (not an actual variable) = sheet_end - radio_in[sheet_output]*sheet_increment;

The trick that Bryan has done to solve this, is to calculate sheet_increment based on the two boundaries:

sheet_increment = (sheet_end - sheet_fully_in)/100.0;

Also, when the RC is turned off, it drops the radio_in[sheet_output] below the sheet_fully_in value; this is UNSAFE. So, we cap out the radio_in[sheet_output] so that it doesn't fall below sheet_fully_in.

Daniel, your job is to clean up this new fix, and verify the code. It is also important that you can read and understand the code.